python/bin/telegram_manager_bot/pyproject.toml
[tool.poetry]
name = "telegram_manager_bot"
version = "0.1.0"
description = ""
authors = ["Nikolay Vasilev <vasilev.nk@phystech.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
poetry-grpc-plugin = "^0.1.1"
proto = {path = "../../lib/proto"}
aiogram = "^2.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
